Explicit types so runtime functions can be detected

+++ contracts/PerpsV2MarketProxyable.sol
@@ -96,7 +96,7 @@ contract PerpsV2MarketProxyable is PerpsV2MarketBase, Proxyable {
         int funding = _nextFundingEntry(price);
         marketState.pushFundingSequence(int128(funding));
         marketState.setFundingLastRecomputed(uint32(block.timestamp));
-        marketState.setFundingRate(fundingRate);
+        marketState.setFundingRate(int128(fundingRate));
 
         emitFundingRecomputed(funding, sequenceLengthBefore, marketState.fundingLastRecomputed());
